pyproject.toml
[project]
name = "enterprise-rag"
version = "0.1.0"
description = "Enterprise-grade Agentic RAG on Azure portfolio project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.7,<3.0",
  "pydantic-settings>=2.3,<3.0",
  "azure-identity>=1.16,<2.0",
  "azure-keyvault-secrets>=4.8,<5.0",
  "azure-search-documents>=11.6,<12.0",
  "openai>=1.35,<2.0",
  "azure-ai-projects>=1.0,<2.0",
  "azure-ai-documentintelligence>=1.0,<2.0",
  "azure-ai-agents>=1.1,<2.0",
  "faiss-cpu>=1.8,<2.0; platform_system != 'Windows'",
  "langgraph>=0.2,<1.0",
  "langchain-openai>=0.1,<1.0",
  "langchain-core>=0.2,<1.0",
  "opentelemetry-api>=1.25,<2.0",
  "opentelemetry-sdk>=1.25,<2.0",
  "azure-monitor-opentelemetry-exporter>=1.0.0b1,<2.0",
  "azure-ai-contentsafety>=1.0,<2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=9.0.3,<10.0",
  "ruff>=0.4.0,<1.0",
  "mypy>=1.10,<2.0",
  "bandit>=1.7,<2.0",
  "pip-audit>=2.8.0,<3.0",
  "pre-commit>=3.7.0,<4.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.mypy]
python_version = "3.11"
strict = true
packages = ["guardrails", "config", "azure_clients", "cost_model", "foundry", "observability"]
